Added tasks page as the default with a stream page.
